# Conflicts:
#	.github/workflows/ci.yml
#	.github/workflows/commitlint.yml
#	.github/workflows/guides.yml
#	.github/workflows/release.yml
#	CHANGELOG.md
#	composer.json
#	src/Mcp/composer.json
#	tests/Functional/JsonLd/SerializableItemDataProviderTest.php

+++ src/Hydra/Collection.php
@@ -31,7 +31,7 @@ class Collection
     #[StreamedName('@type')]
     public string $type = 'Collection';
 
-    public float $totalItems;
+    public int $totalItems;
 
     public ?IriTemplate $search = null;
     public ?PartialCollectionView $view = null;
